--- web3swift/Web3/Classes/Web3+Infura.swift
@@ -1,21 +1,0 @@
-//  web3swift
-//
-//  Created by Alex Vlasov.
-//  Copyright © 2018 Alex Vlasov. All rights reserved.
-//
-
-import Foundation
-import BigInt
-
-/// Custom Web3 HTTP provider of Infura nodes.
-public final class InfuraProvider: Web3HttpProvider {
-    public init?(_ net:Networks, accessToken token: String? = nil, keystoreManager manager: KeystoreManager? = nil) {
-        var requestURLstring = "https://" + net.name + ".infura.io/"
-        if token != nil {
-            requestURLstring = requestURLstring + token!
-        }
-        let providerURL = URL(string: requestURLstring)
-        super.init(providerURL!, network: net, keystoreManager: manager)
-    }
-}
-
